requirements.txt
et_xmlfile==2.0.0
numpy==2.4.2
openpyxl==3.1.5
pandas==3.0.0
pdf2image==1.17.0
pillow==12.1.0
python-dateutil==2.9.0.post0
six==1.17.0
